backend/requirements.txt
torch>=2.1
torchvision>=0.16
fastapi>=0.110
uvicorn[standard]>=0.23
python-multipart>=0.0.9
pillow>=10.0
numpy>=1.26
